+++ roboToe/src/main/java/frc/robot/commands/FourBarRocketLvlOneGroup.java
@@ -6,7 +6,7 @@ import frc.robot.Constants;
 
 public class FourBarRocketLvlOneGroup extends CommandGroup {
 
-  public FourBarRocketLvlOneGroup(int setRocketLvlOnePoint) {
+  public FourBarRocketLvlOneGroup(double setRocketLvlOnePoint) {
 
     addSequential(new WristCommand(Constants.wristUpSetPoint));
     addSequential(new FourBarCommand(Constants.setRocketLvlOnePoint));
